+++ build_local.py
@@ -313,7 +313,7 @@ from PyInstaller.utils.hooks import collect_all
 datas = {datas_literal}
 binaries = []
 hiddenimports = ['spacy', 'presidio_analyzer']
-for pkg in ['presidio_analyzer', 'spacy', 'en_core_web_md', 'it_core_news_md']:
+for pkg in ['presidio_analyzer', 'spacy', 'tldextract', 'en_core_web_md', 'it_core_news_md']:
     t = collect_all(pkg)
     datas += t[0]; binaries += t[1]; hiddenimports += t[2]
 
@@ -329,7 +329,7 @@ from PyInstaller.utils.hooks import collect_all
 datas = {datas_literal}
 binaries = []
 hiddenimports = ['spacy', 'presidio_analyzer']
-for pkg in ['presidio_analyzer', 'spacy', 'en_core_web_md', 'it_core_news_md']:
+for pkg in ['presidio_analyzer', 'spacy', 'tldextract', 'en_core_web_md', 'it_core_news_md']:
     t = collect_all(pkg)
     datas += t[0]; binaries += t[1]; hiddenimports += t[2]